requirements.txt
astropy
easydict
matplotlib
natsort
open3d
opencv-python
pandas
scikit-image
scikit-learn
scipy
Shapely
tensorboard
termcolor
torch
numpy
Pillow
tqdm